Create an option module
